fix(money): use Decimal for monetary calculations
Move receipt, payment, and balance money fields to Decimal handling and store new monetary values in stable string form.

Balance calculations now read legacy numeric values through Decimal and round money explicitly, avoiding binary float drift in financial rows.

+++ app/schemas.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field
@@ -66,19 +67,19 @@ class AddParticipantsRequest(BaseModel):
 
 class CreateShareItemRequest(BaseModel):
     user_id: UUID
-    share_value: float = Field(gt=0, le=1)
+    share_value: Decimal = Field(gt=0, le=1)
 
 
 class ShareItem(BaseModel):
     id: UUID
     receipt_item_id: UUID
     user_id: UUID
-    share_value: float = Field(gt=0, le=1)
+    share_value: Decimal = Field(gt=0, le=1)
 
 
 class CreateReceiptItemRequest(BaseModel):
     name: str = ""
-    cost: float = Field(gt=0)
+    cost: Decimal = Field(gt=0)
     share_items: list[CreateShareItemRequest] = Field(min_length=1)
 
 
@@ -86,20 +87,20 @@ class ReceiptItem(BaseModel):
     id: UUID
     receipt_id: UUID
     name: str = ""
-    cost: float
+    cost: Decimal
     share_items: list[UUID]
 
 
 class CreateReceiptRequest(BaseModel):
     payer_id: UUID
     title: str = ""
-    total_amount: float = Field(gt=0)
+    total_amount: Decimal = Field(gt=0)
     items: list[CreateReceiptItemRequest] = Field(min_length=1)
 
 
 class UpdateReceiptRequest(BaseModel):
     title: str | None = None
-    total_amount: float | None = Field(default=None, gt=0)
+    total_amount: Decimal | None = Field(default=None, gt=0)
     items: list[CreateReceiptItemRequest] | None = None
 
 
@@ -108,7 +109,7 @@ class Receipt(BaseModel):
     event_id: UUID
     payer_id: UUID
     title: str = ""
-    total_amount: float
+    total_amount: Decimal
     created_at: datetime
     updated_at: datetime
     items: list[ReceiptItem]
@@ -126,7 +127,7 @@ class ReceiptImagePresignedUrlResponse(BaseModel):
 class PaymentCreate(BaseModel):
     sender_id: UUID
     receiver_id: UUID
-    amount: float = Field(gt=0)
+    amount: Decimal = Field(gt=0)
 
 
 class PaymentUpdate(BaseModel):
@@ -138,7 +139,7 @@ class Payment(BaseModel):
     event_id: UUID
     sender_id: UUID
     receiver_id: UUID
-    amount: float
+    amount: Decimal
     confirmed: bool
     created_at: datetime
 
@@ -147,4 +148,4 @@ class EventBalance(BaseModel):
     event_id: UUID
     debitor_id: UUID
     creditor_id: UUID
-    amount: float
+    amount: Decimal
